Add app to accounts and css fix
