+++ aMovieDatabase/accounts/models.py
@@ -1,3 +1,5 @@
 from django.db import models
 
 # Create your models here.
+from django.db import models
+from django.contrib.auth.models import AbstractUser
