+++ aMovieDatabase/accounts/static/style.css
@@ -29,7 +29,7 @@
 .form-control {
 	height: 40px;
 	box-shadow: none;
-	color: #969fa4;
+	color: black;
 }
 .form-control:focus {
 	border-color: #5cb85c;
