+++ aMovieDatabase/accounts/templates/login.html
@@ -1,5 +1,7 @@
 {% extends "base.html" %}
  {% load static %}
+
+
 {% block content %}
 <head>
       <link rel="stylesheet" type="text/css" href="{%static 'style.css' %}">
@@ -33,6 +35,7 @@
         </div>
     </form>
     <p class="text-center"><a href="/register">Create an Account</a></p>
+</div>
 </div>
 </body>
 </html>
