+++ aMovieDatabase/core/urls.py
@@ -1,4 +1,3 @@
-from django.urls import path
 
 from . import views
 from django.urls import path
@@ -7,7 +6,4 @@ from django.contrib.auth import views as auth_views
 
 urlpatterns = [
     path('', views.home, name='home'),
-    path('register/', views.registration, name='register'),
-    path('login/', views.login_view, name='login'),
-    path('logout/', auth_views.LogoutView.as_view(template_name='home.html'), name='logout'),
 ]
